nlm init

+++ cortex/imaging/fourier.c
@@ -12,6 +12,7 @@ cooley-tukey
 
 // https://www.originlab.com/doc/Origin-Help/FFT2-Algorithm
 void fft(struct Image input, struct CImage output){
+	
 	uint8_t v, u;             // wave components
 	uint8_t y, x;             // spatial component
 	float modulator;
@@ -23,8 +24,8 @@ void fft(struct Image input, struct CImage output){
 			for(y = 0; y < input.size; ++y){
 				for(x = 0; x < input.size; ++x){
 					modulator = -2.0 * PI * ((1.0 * u * x + 1.0 * v * y) / input.size);
-					output.re[v][u] += input[y][x] * cos(modulator);
-					output.im[v][u] += input[y][x] * sin(modulator);
+					output.re[v][u] += input[y][x] * cosf(modulator);
+					output.im[v][u] += input[y][x] * sinf(modulator);
 				}
 			}
 		}
@@ -33,14 +34,15 @@ void fft(struct Image input, struct CImage output){
 	// normalize by avg
 	for(v = 0; v < input.size; ++v){
 		for(u = 0; u < input.size; ++u){
-				output.re[v][u] /= norm;
-				output.im[v][u] /= norm;
+			output.re[v][u] /= norm;
+			output.im[v][u] /= norm;
 		}
 	}
 }
 
 // https://www.originlab.com/doc/Origin-Help/InverseFFT2-Algorithm
 void ifft(struct CImage input, struct Image output){
+	
 	uint8_t v, u;             // wave components
 	uint8_t y, x;             // spatial component
 	float modulator;
@@ -52,7 +54,7 @@ void ifft(struct CImage input, struct Image output){
 			for(v = 0; v < input.size; ++v){
 				for(u = 0; u < input.size; ++u){
 					modulator = 2.0 * PI * ((1.0 * u * x + 1.0 * v * y) / input.size);
-					output.data[y][x] += input.re[v][u] * cos(modulator) + input.im[v][u] * sin(modulator);
+					output.data[y][x] += input.re[v][u] * cosf(modulator) + input.im[v][u] * sinf(modulator);
 				}
 			}
 		}
